+++ gather/prometheus.go
@@ -2,6 +2,7 @@ package gather
 
 import (
 	"context"
+	"crypto/tls"
 	"fmt"
 	"io"
 	"math"
@@ -17,11 +18,31 @@ import (
 
 // prometheusScraper handles parsing prometheus metrics.
 // implements Scraper interfaces.
-type prometheusScraper struct{}
+type prometheusScraper struct {
+	insecureHttp *http.Client
+}
+
+// newPrometheusScraper create a new prometheusScraper.
+func newPrometheusScraper() *prometheusScraper {
+	customTransport := http.DefaultTransport.(*http.Transport).Clone()
+	customTransport.TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
+	client := &http.Client{Transport: customTransport}
+
+	return &prometheusScraper{insecureHttp: client}
+}
 
 // Gather parse metrics from a scraper target url.
 func (p *prometheusScraper) Gather(ctx context.Context, target influxdb.ScraperTarget) (collected MetricsCollection, err error) {
-	resp, err := http.Get(target.URL)
+	var (
+		resp *http.Response
+	)
+
+	if target.AllowInsecure {
+		resp, err = p.insecureHttp.Get(target.URL)
+	} else {
+		resp, err = http.Get(target.URL)
+	}
+
 	if err != nil {
 		return collected, err
 	}